==> quality_check/08-gene_lengths.Rmd <==
---
title: "Zebrafish gene lengths"
output: 
  html_notebook:
    toc: TRUE
    toc_float: TRUE
author: J. Taroni for CCDL
date: 2019
---

In `7-technology_diff_exp`, we observed that genes that are shorter tend to have
lower expression values in RNA-seq data in the test compendium.

Here, we're interested in if short genes tend to be unobserved (e.g., have
counts/count-scale data that are zero) in the RNA-seq data used in the test 
compendium.
We'll go back to the `lengthScaledTPM.tsv` that has not yet been through
the aggregation process to explore this.
Note that we will only be using RNA-seq files that were used in an earlier 
analysis of imputation performance.
These will be ~400 RNA-seq samples that were included in a smaller
test compendium and thus, for later analyses in this notebook we'll use the
smaller test compendium.

## Set up

```{r}
`%>%` <- dplyr::`%>%`
# we'll need ggfortify to do the PCA plots
library(ggfortify)
```

### Custom functions

These are a series of wrapper functions for plotting, as we typically make
plots for _all genes_ and then a subset of genes (such as those that are
in the compendium).
They generally expect certain objects to be in the global environment;
see the documentation for the individual functions.

```{r}
zero_proportion_point_wrapper <- function(df, plot.subtitle) {
  # Make the point plots (with geom_smooth) for the proportion of samples that
  # have a value of 0 for a gene vs. that gene's length
  # 
  # Args:
  #   df: a data.frame that includes 'gene_length' and 'Zero_proportion'
  #   plot.subtitle: subtitle for the plot; character
  # 
  # Returns:
  #   a ggplot as described above
  
  if(!all(c("gene_length" %in% colnames(df), 
            "Zero_proportion" %in% colnames(df)))) {
    stop("df must have colnames 'gene_length' and 'Zero_proportion'")  
  }
    
  df %>%
    ggplot(aes(x = gene_length, y = Zero_proportion)) +
    geom_point(alpha = 0.2) +
    geom_smooth() +
    theme_bw() +
    labs(x = "gene length", y = "proportion of samples with zero count",
         subtitle = plot.subtitle, title = "Gene length and observations")
}
```

```{r}
pca_plot_wrapper <- function(prcomp_results, plot.title) {
  # given the output of prcomp, make a scatterplot of PC1 and PC2 with 
  # ggfortify::autoplot where the samples are colored by technology (e.g.,
  # microarray, RNA-seq)
  # requires technology_df to be in the global environment
  # Args:
  #  prcomp_results: output of prcomp, run on a transposed expression matrix
  #  plot.title: title of the plot; character
  #
  # Returns
  #   ggplot as described above
  autoplot(prcomp_results, data = technology_df, alpha = 0.5, 
           colour = 'compendium_technology_labels', scale = 0) +
    scale_color_manual(labels = c("MICROARRAY", "RNA-SEQ"), 
                       values = c("#E69F00", "#56B4E9")) +
    theme_bw() +
    ggtitle(plot.title) +
    theme(plot.title = element_text(hjust = 0.5, face = "bold")) +
    guides(colour = guide_legend(title = "technology"))
}
```

```{r}
pca_pairs_plot_wrapper <- function(prcomp_results, plot.title) {
  # given the output of prcomp, make a pairs plot of PC1-5 with where the 
  # samples are colored by technology (e.g., microarray, RNA-seq)
  # requires compendium_technology_labels to be in the global environment
  #
  # Args:
  #  prcomp_results: output of prcomp, run on a transposed expression matrix
  #  plot.title: title of the plot; character
  #
  # Returns
  #   plot as described above
  pairs(prcomp_results$x[, 1:5], 
        col = dplyr::recode(compendium_technology_labels,
                            `MICROARRAY` = "#E69F00",
                            `RNASEQ` = "#56B4E9"),
        main = plot.title)
}
```

### Directories

```{r}
# directory to hold the plots
plot_dir <- "plots"
```

## Read in data 

### `lengthScaledTPM.tsv` files

```{r}
list_of_rnaseq_files <- list.files(file.path("..", "select_imputation_method", 
                                             "data", "rnaseq"),
                                   full.names = TRUE)
```

Read in the files such that they are individual `data.frame` in a list

```{r}
rnaseq_df_list <- lapply(list_of_rnaseq_files, readr::read_tsv)
```

```{r}
# bind all the individual sample data.frame together
rnaseq_df <- dplyr::bind_cols(rnaseq_df_list)
# we end up with a bunch of "extra" gene columns this way, but they should be
# in the same order
all.equal(rnaseq_df$Gene1, rnaseq_df$Gene112)
```

```{r}
# drop the extra gene columns and set the rownames to gene identifiers -- this
# will help with calculating the proportion of zeros, etc.
rnaseq_df <- rnaseq_df %>%
  as.data.frame() %>%
  tibble::column_to_rownames("Gene") %>%
  dplyr::select(-dplyr::contains("Gene"))
# remove the list of individual sample data.frame
rm(rnaseq_df_list)
```

```{r}
# read in gene lengths
gene_lengths_df <- 
  readr::read_tsv(file.path("data", 
                            "Danio_rerio.GRCz11_gene_lengths_exons.tsv"))
```

### Compendium genes, expression matrix, technology labels

We will explore this gene lengths-observations/expression levels relationship in
_all_ genes and also subset to only the genes in the test compendium, so we 
need to pull those gene identifiers out of the test compendium expression
matrix.

```{r}
# read in smaller test compendium and convert to expression matrix
compendium_exprs_file <- file.path("refinebio-data", "DANIO_RERIO", 
                                   "DANIO_RERIO.tsv")
exprs_mat <- readr::read_tsv(compendium_exprs_file, progress = FALSE) %>%
  as.data.frame() %>%
  tibble::column_to_rownames("X1") %>%
  as.matrix()

# derive the technology labels from the sample names
sample_names <- colnames(exprs_mat)
compendium_technology_labels <- rep("MICROARRAY", length(sample_names))
compendium_technology_labels[grep("lengthScaledTPM", sample_names)] <- "RNASEQ"
# get the technology labels as a data.frame for plotting later
technology_df <- as.data.frame(compendium_technology_labels) 

# get the list of compendium genes
compendium_genes <- rownames(exprs_mat)
```

## Gene zero counts and expression levels

### For each gene, how many samples have a count-scale expression value of zero?

```{r}
# calculate the proportion of samples a gene is zero for
gene_zeroes_proportion <- data.frame(
  "Gene" = rownames(rnaseq_df),
  "Zero_proportion" = apply(rnaseq_df, 1, 
                            function(x) sum(x == 0) / length(x))
)
```

```{r}
# bind to lengths, this will be used for plotting the relationship between
# gene length and number of zeroes
zeroes_df <- dplyr::inner_join(gene_lengths_df, gene_zeroes_proportion,
                               by = "Gene")
```

```{r}
# there are some really long genes that will probably make it difficult to
# visualize what is going on
summary(zeroes_df$gene_length)
```

#### All genes

```{r}
# filtering genes based on length to get a clearer picture
zeroes_df %>%
  dplyr::filter(gene_length <= 20000) %>%
  zero_proportion_point_wrapper(plot.subtitle = "All genes")
```

```{r}
ggsave(filename = file.path(plot_dir, 
                            "gene_length_vs_zero_count_all_genes.png"),
       plot = last_plot())
```

#### Compendium genes only

```{r}
# subset just to genes that are in the compendium
zeroes_df %>%
  dplyr::filter(Gene %in% compendium_genes, 
                gene_length <= 20000) %>%
  zero_proportion_point_wrapper(plot.subtitle = "Test compendium genes only")
```

```{r}
ggsave(filename = file.path(plot_dir, 
                            "gene_length_vs_zero_count_compendium_genes.png"),
       plot = last_plot())
```

#### Takeaway

Shorter genes do tend to have a larger proportion of samples where they have
zero values. 
It stands to reason that if we do not _observe_ shorter genes, we can not
correct for length-bias effectively. 
(Recall that `lengthScaledTPM` are calculated using the average tx length
observed across samples.)

## Removing "short" genes from the test species compendium

Does removing shorter genes from the species compendium remove evidence of a
technology bias?

### PCA - all genes

#### PC1 and PC2 for all genes

```{r}
pca_results <- prcomp(t(exprs_mat))
pca_plot_wrapper(prcomp_results = pca_results, 
                 plot.title = "Test Compendium by Technology: All genes")
```

```{r}
ggsave(filename = file.path(plot_dir, 
                            "small_test_compendium_PCA.png"),
       plot = last_plot())
```

#### Pairs plot (PC1-5) for all genes

```{r}
png(file.path(plot_dir, "small_test_compendium_pairs_plot.png"))
pca_pairs_plot_wrapper(prcomp_results = pca_results,
                       plot.title = "Test Compendium by Technology: All genes")
dev.off()
```

### Drop the "short" genes from the expression matrix

```{r}
# what genes are short? using 25th percentile
short_genes <- zeroes_df %>%
  dplyr::filter(gene_length <= quantile(zeroes_df$gene_length, 0.25)) %>%
  dplyr::pull(Gene)
```

```{r}
length_filtered_exprs_mat <- exprs_mat[!rownames(exprs_mat) %in% short_genes, ]
dim(length_filtered_exprs_mat)
```

### PCA with the short genes removed

```{r}
# PCA on the length filtered exprs mat
filtered_pca_results <- prcomp(t(length_filtered_exprs_mat))
```
#### PC1 and PC2 scatterplot

```{r}
pca_plot_wrapper(prcomp_results = filtered_pca_results,
                 plot.title = paste("Test Compendium by Technology:",  
                                    "Short genes removed"))
```

```{r}
ggsave(filename = file.path(plot_dir, 
                            "gene_lengths_test_compendium_PCA.png"),
       plot = last_plot())
```

#### Pairs plot (PC1-5)

```{r}
png(file.path(plot_dir, "gene_lengths_test_compendium_pairs_plot.png"))
pca_pairs_plot_wrapper(prcomp_results = filtered_pca_results,
                       plot.title = paste("Test Compendium by Technology:",  
                                          "Short genes removed"))
dev.off()
```

#### Takeaway

Removing short genes from the test species compendium does not remove 
technology bias, but there are some considerations:

* These data have already been quantile normalized -- if the gene length
contributes to the expression values prior to quantile normalization, this 
bias could possibly "leak into" the quantile normalized values for all the
genes.

## Session Info

```{r}
sessionInfo()
```
